Since this was written before I knew about git, this is a completed project
This is a simple Chess analysis board with many features.  Playing against
Stockfish 15.1 is possible.  Follow the instructions in README to get
Stockfish working.
This was more of an exercise to build my coding skills
than anything else.  It was a lot of fun and I look forward to making
an improved version in the future.